Backend/requirements.txt
fastapi
uvicorn
streamlit
requests
langchain
langgraph
langchain-google-genai
python-dotenv
pydantic